requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
mypy==1.9.0
mypy-extensions==1.0.0
psycopg==3.1.18
python-decouple==3.8
setuptools==68.2.2
sqlparse==0.5.0
typing_extensions==4.11.0
wheel==0.41.2
